+++ src/redux/sagas.js
@@ -3,5 +3,5 @@ import user from "./user/saga";
 import documents from "./documents/saga";
 
 export default function* rootSaga() {
-  yield all([user()]);
+  yield all([user(), documents()]);
 }