+++ src/routes/AppRouter.js
@@ -5,15 +5,15 @@
 import React, { lazy } from "react";
 import { BrowserRouter, Route, Routes } from "react-router-dom";
 
-const Index = lazy(() => import("../pages/Documents/Index"));
-const AddDocument = lazy(() => import("../pages/AddDocument/AddDocument"));
+const Documents = lazy(() => import("../pages/Documents/Index"));
+const AddDocument = lazy(() => import("../pages/AddDocument/Index"));
 const NotFound = lazy(() => import("../pages/NotFound"));
 
 const AppRouter = () => {
   return (
     <BrowserRouter>
       <Routes>
-        <Route path="/" element={<Index />} />
+        <Route path="/" element={<Documents />} />
         <Route path="/subirDocumento/*" element={<AddDocument />} />
         <Route path="*" element={<NotFound />} />
       </Routes>